refactor(discovery): Extract Discovery into a standalone module.
Moved Discovery-related classes, traits, and interfaces from `Tempest\Core` to a new `Tempest\Discovery` namespace. Updated all references, adjusted imports, and added a dedicated composer.json for the new module. This change improves modularity and separates concerns within the codebase.
