+++ composer.json
@@ -63,6 +63,7 @@
         "tempest/core": "self.version",
         "tempest/database": "self.version",
         "tempest/debug": "self.version",
+        "tempest/discovery": "self.version",
         "tempest/event-bus": "self.version",
         "tempest/filesystem": "self.version",
         "tempest/generation": "self.version",
@@ -90,6 +91,7 @@
             "Tempest\\Core\\": "src/Tempest/Core/src",
             "Tempest\\Database\\": "src/Tempest/Database/src",
             "Tempest\\Debug\\": "src/Tempest/Debug/src",
+            "Tempest\\Discovery\\": "src/Tempest/Discovery/src",
             "Tempest\\EventBus\\": "src/Tempest/EventBus/src",
             "Tempest\\Filesystem\\": "src/Tempest/Filesystem/src",
             "Tempest\\Framework\\": "src/Tempest/Framework",
@@ -129,6 +131,7 @@
             "Tempest\\Container\\Tests\\": "src/Tempest/Container/tests",
             "Tempest\\Core\\Tests\\": "src/Tempest/Core/tests",
             "Tempest\\Database\\Tests\\": "src/Tempest/Database/tests",
+            "Tempest\\Discovery\\Tests\\": "src/Tempest/Discovery/tests",
             "Tempest\\EventBus\\Tests\\": "src/Tempest/EventBus/tests",
             "Tempest\\Filesystem\\Tests\\": "src/Tempest/Filesystem/tests",
             "Tempest\\Generation\\Tests\\": "src/Tempest/Generation/tests",
