+++ public/index.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use Tempest\Core\DiscoveryLocation;
+use Tempest\Discovery\DiscoveryLocation;
 use Tempest\Router\HttpApplication;
 
 require_once __DIR__ . '/../vendor/autoload.php';
