+++ src/Tempest/Core/composer.json
@@ -7,6 +7,7 @@
         "php": "^8.4",
         "tempest/container": "dev-main",
         "tempest/support": "dev-main",
+        "tempest/discovery": "dev-main",
         "vlucas/phpdotenv": "^5.6",
         "filp/whoops": "^2.15"
     },
