+++ src/Tempest/Core/src/Kernel/LoadDiscoveryClasses.php
@@ -10,14 +10,14 @@ use RecursiveIteratorIterator;
 use SplFileInfo;
 use Tempest\Cache\DiscoveryCacheStrategy;
 use Tempest\Container\Container;
-use Tempest\Core\DiscoversPath;
-use Tempest\Core\Discovery;
 use Tempest\Core\DiscoveryCache;
 use Tempest\Core\DiscoveryDiscovery;
-use Tempest\Core\DiscoveryItems;
-use Tempest\Core\DiscoveryLocation;
-use Tempest\Core\DoNotDiscover;
 use Tempest\Core\Kernel;
+use Tempest\Discovery\DiscoversPath;
+use Tempest\Discovery\Discovery;
+use Tempest\Discovery\DiscoveryItems;
+use Tempest\Discovery\DiscoveryLocation;
+use Tempest\Discovery\DoNotDiscover;
 use Tempest\Reflection\ClassReflector;
 use Throwable;
