+++ src/Tempest/Core/src/PublishesFiles.php
@@ -8,6 +8,7 @@ use Closure;
 use Tempest\Console\Exceptions\ConsoleException;
 use Tempest\Console\HasConsole;
 use Tempest\Container\Inject;
+use Tempest\Discovery\DoNotDiscover;
 use Tempest\Generation\ClassManipulator;
 use Tempest\Generation\DataObjects\StubFile;
 use Tempest\Generation\Enums\StubFileType;
