+++ src/Tempest/Database/src/MigrationDiscovery.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Tempest\Database;
 
-use Tempest\Core\DiscoversPath;
-use Tempest\Core\Discovery;
-use Tempest\Core\DiscoveryLocation;
-use Tempest\Core\IsDiscovery;
+use Tempest\Discovery\DiscoversPath;
+use Tempest\Discovery\Discovery;
+use Tempest\Discovery\DiscoveryLocation;
+use Tempest\Discovery\IsDiscovery;
 use Tempest\Reflection\ClassReflector;
 
 final class MigrationDiscovery implements Discovery, DiscoversPath
