+++ src/Tempest/EventBus/src/EventBusDiscovery.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Tempest\EventBus;
 
 use BackedEnum;
-use Tempest\Core\Discovery;
-use Tempest\Core\DiscoveryLocation;
-use Tempest\Core\IsDiscovery;
+use Tempest\Discovery\Discovery;
+use Tempest\Discovery\DiscoveryLocation;
+use Tempest\Discovery\IsDiscovery;
 use Tempest\Reflection\ClassReflector;
 use Tempest\Reflection\TypeReflector;
 use UnitEnum;
