+++ src/Tempest/View/src/ViewComponentDiscovery.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Tempest\View;
 
-use Tempest\Core\DiscoversPath;
-use Tempest\Core\Discovery;
-use Tempest\Core\DiscoveryLocation;
-use Tempest\Core\IsDiscovery;
+use Tempest\Discovery\DiscoversPath;
+use Tempest\Discovery\Discovery;
+use Tempest\Discovery\DiscoveryLocation;
+use Tempest\Discovery\IsDiscovery;
 use Tempest\Reflection\ClassReflector;
 use Tempest\View\Components\AnonymousViewComponent;
