+++ tests/Fixtures/Discovery/HiddenDatabaseMigration.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Tests\Tempest\Fixtures\Discovery;
 
-use Tempest\Core\DoNotDiscover;
 use Tempest\Database\DatabaseMigration;
 use Tempest\Database\QueryStatement;
+use Tempest\Discovery\DoNotDiscover;
 
 #[DoNotDiscover]
 final class HiddenDatabaseMigration implements DatabaseMigration
