+++ tests/Fixtures/TestInstallerClass.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\Tempest\Fixtures;
 
-use Tempest\Core\DoNotDiscover;
+use Tempest\Discovery\DoNotDiscover;
 
 #[DoNotDiscover]
 final readonly class TestInstallerClass
