+++ tests/Integration/Core/KernelTest.php
@@ -6,8 +6,8 @@ namespace Tests\Tempest\Integration\Core;
 
 use PHPUnit\Framework\TestCase;
 use Tempest\Container\Container;
-use Tempest\Core\DiscoveryLocation;
 use Tempest\Core\Kernel;
+use Tempest\Discovery\DiscoveryLocation;
 use Tests\Tempest\Fixtures\TestDependency;
 
 /**
