+++ tests/Integration/FrameworkIntegrationTestCase.php
@@ -12,11 +12,11 @@ use Tempest\Console\OutputBuffer;
 use Tempest\Console\Testing\ConsoleTester;
 use Tempest\Core\AppConfig;
 use Tempest\Core\Application;
-use Tempest\Core\DiscoveryLocation;
 use Tempest\Core\ShellExecutor;
 use Tempest\Core\ShellExecutors\NullShellExecutor;
 use Tempest\Database\DatabaseConfig;
 use Tempest\Database\Migrations\MigrationManager;
+use Tempest\Discovery\DiscoveryLocation;
 use Tempest\Framework\Testing\IntegrationTest;
 use Tempest\Router\HttpApplication;
 use Tempest\View\Components\AnonymousViewComponent;
